[Refactor]기존 Repository LoginRepository로 변경 후 Domain과 Data 패키지에 분리
