+++ app/src/main/java/com/doctor/yumyum/data/repository/LoginRepositoryImpl.kt
@@ -3,23 +3,14 @@ package com.doctor.yumyum.data.repository
 import com.doctor.yumyum.data.local.LocalDataSourceImpl
 import com.doctor.yumyum.data.model.signUpModel
 import com.doctor.yumyum.data.remote.RemoteDataSourceImpl
+import com.doctor.yumyum.domain.repository.LoginRepository
 import okhttp3.ResponseBody
 import retrofit2.Call
 import retrofit2.Response
 
-interface Repository {
-    val localDataSource: LocalDataSourceImpl
-    val remoteDataSource: RemoteDataSourceImpl
 
-    // 로그인 토큰
-    fun getLoginToken(): String?
-    fun setLoginToken(loginToken: String)
 
-    // 회원가입
-    fun postAuthCreation(signUpModel: signUpModel): retrofit2.Call<ResponseBody>
-}
-
-class RepositoryImpl : Repository {
+class LoginRepositoryImpl : LoginRepository {
     override val localDataSource: LocalDataSourceImpl
         get() = LocalDataSourceImpl()
     override val remoteDataSource: RemoteDataSourceImpl
