--- app/src/main/java/com/doctor/yumyum/domain/repository/README.md
@@ -1,1 +1,0 @@
-## repository interface
