+++ app/src/main/java/com/doctor/yumyum/presentation/ui/splash/SplashViewModel.kt
@@ -1,9 +1,9 @@
 package com.doctor.yumyum.presentation.ui.splash
 
 import com.doctor.yumyum.common.base.BaseViewModel
-import com.doctor.yumyum.data.repository.RepositoryImpl
+import com.doctor.yumyum.data.repository.LoginRepositoryImpl
 
 class SplashViewModel : BaseViewModel() {
-    private val repository: RepositoryImpl = RepositoryImpl()
+    private val repository: LoginRepositoryImpl = LoginRepositoryImpl()
     val loginToken:String? = repository.getLoginToken()
 }
